rentpe/requirements.txt
dj-database-url == 2.0.0
Django == 4.2.3
django-cors-headers == 4.2.0
djangorestframework == 3.14.0
djangorestframework-simplejwt == 5.2.2
gunicorn == 21.2.0
psycopg2 == 2.9.6
PyJWT == 2.8.0
cloudinary == 1.33.0